add more docs and rewrite hl.scan as class instance
